(PC-23190)[API] chore: extract patch product test
